requirements.txt
line-bot-sdk
flask
gensim
jieba